backend/requirements.txt
fastapi
uvicorn
torch
httpx
google-generativeai
python-dotenv
numpy
pandas
timezonefinder
